gru: enable pp1500 and pp3000 rails as soon as possible
The idea is that they stay low unless we know that we booted from SPI
flash. As this code runs in SPI flash - it is ok to turn these rails
on as soon as possible, and pp3000 rail it is essential for UART to
work.

Proto2 Kevin and Gru designs are going to be using these pins to
control these rails. Proto1 Kevin had those GPIO pins routed to two
chip enable signals, it is save to assert them high.

BRANCH=none
BUG=none
TEST=proto1 still boots (which does not prove much)

Change-Id: I31bb03334ad9e3aa57db726fb43dec85014a3f05
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>